requirements.txt
datasets==2.9.0
gradio==3.31.0
pandas==1.5.3
pinecone_client==2.2.1
pinecone-client[grpc]
sentence_transformers==2.2.2
torch==2.0.1
tqdm==4.64.1
